desktop_app_requirements.txt
pillow==10.0.1
pystray==0.19.4
requests==2.31.0
nrgpy==0.7.0
data_email_client
pandas==2.3.2
numpy==2.3.2
